Replaces Patriot STR with AN/FPS-117 EWR
Is there a reason to use the Patriot STR instead of the AN/FPS-117 EWR? The latter has a 360 scan and can be given EWR tasking in the ME.  

Will this affect Skynet functionality? I don't use Skynet but don't want to break it for those that do. Note-I've proposed the same change in the Bluefor Modern faction.